Edge/modules/lightsModulePython/requirements.txt
azure-iot-device~=2.0.0
RPi.GPIO